+++ src/applications/ivc-champva/10-10D/pages/MissingFileConsentPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import MissingFileOverview from '../components/File/MissingFileOverview';
+import MissingFileOverview from '../../shared/components/fileUploads/MissingFileOverview';
 
 export function MissingFileConsentPage(props) {
   const OverviewComp = MissingFileOverview({
