+++ src/applications/ivc-champva/shared/components/fileUploads/upload.js
@@ -2,7 +2,11 @@ import environment from 'platform/utilities/environment';
 import fileUploadUI from 'platform/forms-system/src/js/definitions/file';
 import { focusElement } from 'platform/utilities/ui';
 import VaSelectField from 'platform/forms-system/src/js/web-component-fields/VaSelectField';
-import { fileTypes, maxSize, minSize } from '../../config/attachments';
+import {
+  fileTypes,
+  maxSize,
+  minSize,
+} from '../../../10-10D/config/attachments';
 
 // Modified version of the file upload from applications/appeals/995
